Rename module `util-base` -> `util-common-base`.
Because it's specifically the base for `util-common`.
