+++ local-modules/typecheck/package.json
@@ -6,6 +6,6 @@
   "dependencies": {
     "object-inspect": "^1.2.1",
 
-    "util-base": "local"
+    "util-common-base": "local"
   }
 }
