+++ local-modules/util-common-base/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "util-base",
+  "name": "util-common-base",
   "version": "1.0.0",
   "main": "main.js"
 }
