+++ local-modules/util-common/DataUtil.js
@@ -3,7 +3,7 @@
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
 import { TArray, TInt, TString } from 'typecheck';
-import { ObjectUtil } from 'util-base';
+import { ObjectUtil } from 'util-common-base';
 
 /**
  * "Data value" helper utilities. A "data value" is defined as any JavaScript
